requirements.txt
discord.py>=2.3
anthropic>=0.40
groq>=1.0
python-dotenv>=1.0
aiohttp>=3.9
